# Optimize Filter.andThen