Problem

The definition of `andThen` which accepts a `Filter` parameter is the
most common usage for composing filters and a service, e.g.,

    filterA andThen filterB andThen service

The resulting Service allocates an additional Service per filter for
every request. Merely associating `andThen` to the right avoids the
allocations.

    filterA andThen (filterB andThen service)

This is possible because the right-associated pattern invokes the
definition of `andThen` which accepts a Service parameter. However, it
is not possible to do this if the user wants to build a filter chain
without a service on hand.

Solution

Introduce a hidden type, `AndThen`, to wrap a continuation, which, when
ready to build the Service, we can  use to magically re-associate the
`andThen`s to the most efficient pattern: to the right. The continuation
represents the eventual Service, applied to the filter chain to build a
Service, and therefore has the same type as the `andThen` which takes a
Service parameter: `Service[ReqOut, RepIn] => Service[ReqIn, RepOut]`.

Benchmark

Old

                                       (numAndThens)  Mode  Cnt     Score     Error   Units
    andThenFilter                                  1  avgt   10    10.162 ±   0.077   ns/op
    andThenFilter:·gc.alloc.rate.norm              1  avgt   10    ≈ 10⁻⁵              B/op
    andThenFilter                                 10  avgt   10    95.834 ±   1.872   ns/op
    andThenFilter:·gc.alloc.rate.norm             10  avgt   10   240.000 ±   0.001    B/op
    andThenFilter                                 20  avgt   10   212.385 ±   3.959   ns/op
    andThenFilter:·gc.alloc.rate.norm             20  avgt   10   480.000 ±   0.001    B/op


New

                                       (numAndThens)  Mode  Cnt   Score    Error   Units
    andThenFilter                                  1  avgt   10  11.199 ±  0.181   ns/op
    andThenFilter:·gc.alloc.rate.norm              1  avgt   10  ≈ 10⁻⁵             B/op
    andThenFilter                                 10  avgt   10  47.235 ±  0.507   ns/op
    andThenFilter:·gc.alloc.rate.norm             10  avgt   10  ≈ 10⁻⁴             B/op
    andThenFilter                                 20  avgt   10  85.747 ±  1.062   ns/op
    andThenFilter:·gc.alloc.rate.norm             20  avgt   10  ≈ 10⁻⁴             B/op