@luciferous 2015-10-19
Passed:

    cd finagle; ./sbt 'project finagle-core' test
    cd finagle; ./sbt 'project finagle-http' test
    cd twitter-server; ./sbt test

Benchmark looks too good to be true, though. Let me know how it goes.