@dschobel 2015-10-19
LGTM, very nice!