@vkostyukov 2015-10-26
@luciferous I'm pulling this internally. Will close this PR once it's merged. Thanks! This is a solid work.