@vkostyukov 2015-10-27
Woot! We've merged this internally and it's already on [Github](https://github.com/twitter/finagle/commit/47b27438758782fcfbf37f2bd9947d7a49709b85). Thanks a lot @luciferous!